Simplify further registrations
