+++ src/Win2Mqtt.SystemActions/ServiceCollectionExtensions.cs
@@ -4,12 +4,7 @@ namespace Win2Mqtt.SystemActions
 {
     public static class ServiceCollectionExtensions
     {
-        public static IServiceCollection AddSystemActions(this IServiceCollection services) =>
-            services.Scan(scan => scan
-            .FromAssemblies([typeof(ServiceCollectionExtensions).Assembly])
-            .AddClasses(classes => classes.AssignableTo<ISystemActionWrapper>())
-			.AsSelf()
-            .AsImplementedInterfaces()
-            .WithSingletonLifetime());
+        public static IServiceCollection AddSystemActions(this IServiceCollection services) => services
+            .AddSystemActionsFromAssembly(typeof(ServiceCollectionExtensions).Assembly);
     }
 }
